Stage #1: Welcome
Description

Hypercar Ltd is a not-so-famous car manufacturer. Last year was a very prospective one for the company: they doubled the profit and sold about 10k cars! Every Hypercar sometimes needs to be fixed or checked-up, so the company also opened several service centers. Soon managers noticed that the operational process is not optimal for customers, so they need a web-application for a queue management system and they want you to implement this service.

You start working on the "hypercar" Django project with the "tickets" application. You will make a server that creates new tickets for customers and ensures the correct order of the queue.

Objectives

At first, we want to greet all the customers, so please implement the welcome page for an application with a particular message.

To emphasize the text create an <h2> element with the content "Welcome to the Hypercar service!":

<h2>Welcome to the Hypercar Service!</h2>
The project will have several URLs for different tasks, so please make this page available at "<base_url>/welcome", where <base_url> is the host. If you start the application on your computer with python manage.py runserver command, the host is localhost:8000.

If the server returns the page at the address "http://localhost:8000/welcome", you can be sure it works fine!