Stage #3: Electronic Queue
Description

Every day at the center there are two or three hard-working mechanics and they take clients in the order they are coming. Sometimes the line becomes really long when customers drive in at the same time. Usually, the diagnostic test takes 30 minutes, inflating the tires 5 minutes and changing the oil about 2 minutes.

The company conducts a survey and come to the conclusion: the longer the service takes, the longer the client agrees to wait. That's why a new algorithm of processing should be implemented:

As long as there are clients in the line waiting to change oil, they are served first;
The ones who want to inflate tires are served next
The clients who come to get diagnostics go last
To distinguish one ticket from another we should choose a number for it. Starting from 1 and getting next in increasing order is the expected solution from you. Not all the clients agree to wait long, so give them the estimated time before their turn. Do not forget that you have a particular order of processing the line.

Objectives

Your task is to create a handler that gives tickets to the clients. The handler should serve GET requests on URLs "<base_url>/get_ticket/change_oil", "<base_url>/get_ticket/inflate_tires", "<base_url>/get_ticket/diagnostic" where <base_url>  is the host.

On a page there should be a <div> element with the content "Your number is <ticket_number>", where the <ticket_number> is an id of a ticket. To calculate estimated time, sum up the time of processing all the tickets before the given one according to the algorithm above. The estimated time should be in a <div> element with content "Please wait around <minutes_to_wait> minutes" where <minutes_to_wait> is the number of estimated minutes:

<div>Your number is 7</div>
<div>Please wait around 49 minutes</div>
To store the data, you may use Python dictionaries and lists. For example, you can create a dictionary line_of_cars with the keys matching service names and values having a list of customers' tickets.