Stage #4: Operator Menu
Description

The service you're making is almost ready for customers, but you should also make it handy for operators who process the orders. They want to see the current length of each queue. If lengths are 3, 4 and 1, the content of each paragraph will be:

Change oil queue: 3
Inflate tires queue: 4
Get diagnostic queue: 1
Another action they want to make is to choose when to process the next client. The simple button for this operation will serve great.

Objectives

Please add three <div> elements with content from the list above. The last element on a page should be a <form> with the method attribute equal to "post".

The form should contain a <button> element with the type attribute equal to "submit". The label of the button is "Process next":

<div>Change oil queue: 3</div>
<div>Inflate tires queue: 4</div>
<div>Get diagnostic queue: 1</div>

<form method="post">{% csrf_token %}
  <button type="submit">Process next</button>
</form>
Do not implement the handler for the form action: you'll do it at the next step.

The page with the operator menu should be available at "<base_url>/processing", where <base_url> is the host.