Stage #5: Serving Next Client
Description

When the next client's turn is coming up, the operator should process it pushing the button. It happens when one of the mechanics has finished the work and able to serve the next client. You cannot define in the application the exact time, so leave the task to the operator.

When the "Process next" button is pushed the application chooses the next client according to the algorithm's priority:

Until the line of customers to change the oil is not empty, they served first
If "Change oil" queue is empty the line of customers to inflate tires served next
Clients who want to get diagnostics served last.
Immediately the number of chosen ticket appears on the screen for all the customers and the length of the appropriate queue become one client less. You need not save the ticket further, the only purpose of it now is to show the number on the screen.

Objectives

Your task is to make a POST handler for processing the next ticket. After the request is received, remove the appropriate ticket from the queue. The handler should process POST requests at "<base_url>/processing", where <base_url> is the host.

Remember that you should have another handler to show information about the next ticket.

Add "Next ticket #<number_of_ticket>" to the <div> element to the page that will be displayed on the screen for all the clients:

<div>Next ticket #7</div>
If there are no tickets to process, the content of the element should be "Waiting for the next client":

<div>Waiting for the next client</div>
The page with the next ticket should be available at "<base_url>/next", where <base_url> is the host.

Be cautious, URLs in the urls.py module should not contain slash in the end.